+++ apps/webapp/package.json
@@ -150,7 +150,7 @@
     "simple-oauth2": "^5.0.0",
     "simplur": "^3.0.1",
     "slug": "^6.0.0",
-    "socket.io": "^4.7.4",
+    "socket.io": "4.7.4",
     "socket.io-adapter": "^2.5.4",
     "sonner": "^1.0.3",
     "sqs-consumer": "^7.4.0",
