+++ package.json
@@ -68,7 +68,8 @@
   "pnpm": {
     "patchedDependencies": {
       "@changesets/assemble-release-plan@5.2.4": "patches/@changesets__assemble-release-plan@5.2.4.patch",
-      "tsup@8.0.1": "patches/tsup@8.0.1.patch"
+      "tsup@8.0.1": "patches/tsup@8.0.1.patch",
+      "engine.io-parser@5.2.2": "patches/engine.io-parser@5.2.2.patch"
     }
   }
 }
