+++ packages/core-apps/package.json
@@ -28,7 +28,8 @@
     "@trigger.dev/core": "workspace:*",
     "@trigger.dev/tsconfig": "workspace:*",
     "@types/node": "18",
-    "typescript": "^5.3.0"
+    "typescript": "^5.3.0",
+    "socket.io-client": "^4.7.4"
   },
   "engines": {
     "node": ">=18.0.0"
