+++ packages/core/package.json
@@ -37,6 +37,70 @@
       "require": "./dist/v3/otel/index.js",
       "types": "./dist/v3/otel/index.d.ts"
     },
+    "./v3/zodMessageHandler": {
+      "import": {
+        "types": "./dist/v3/zodMessageHandler.d.mts",
+        "default": "./dist/v3/zodMessageHandler.mjs"
+      },
+      "require": "./dist/v3/zodMessageHandler.js",
+      "types": "./dist/v3/zodMessageHandler.d.ts"
+    },
+    "./v3/zodNamespace": {
+      "import": {
+        "types": "./dist/v3/zodNamespace.d.mts",
+        "default": "./dist/v3/zodNamespace.mjs"
+      },
+      "require": "./dist/v3/zodNamespace.js",
+      "types": "./dist/v3/zodNamespace.d.ts"
+    },
+    "./v3/zodSocket": {
+      "import": {
+        "types": "./dist/v3/zodSocket.d.mts",
+        "default": "./dist/v3/zodSocket.mjs"
+      },
+      "require": "./dist/v3/zodSocket.js",
+      "types": "./dist/v3/zodSocket.d.ts"
+    },
+    "./v3/zodIpc": {
+      "import": {
+        "types": "./dist/v3/zodIpc.d.mts",
+        "default": "./dist/v3/zodIpc.mjs"
+      },
+      "require": "./dist/v3/zodIpc.js",
+      "types": "./dist/v3/zodIpc.d.ts"
+    },
+    "./v3/utils/structuredLogger": {
+      "import": {
+        "types": "./dist/v3/utils/structuredLogger.d.mts",
+        "default": "./dist/v3/utils/structuredLogger.mjs"
+      },
+      "require": "./dist/v3/utils/structuredLogger.js",
+      "types": "./dist/v3/utils/structuredLogger.d.ts"
+    },
+    "./v3/dev": {
+      "import": {
+        "types": "./dist/v3/dev/index.d.mts",
+        "default": "./dist/v3/dev/index.mjs"
+      },
+      "require": "./dist/v3/dev/index.js",
+      "types": "./dist/v3/dev/index.d.ts"
+    },
+    "./v3/prod": {
+      "import": {
+        "types": "./dist/v3/prod/index.d.mts",
+        "default": "./dist/v3/prod/index.mjs"
+      },
+      "require": "./dist/v3/prod/index.js",
+      "types": "./dist/v3/prod/index.d.ts"
+    },
+    "./v3/workers": {
+      "import": {
+        "types": "./dist/v3/workers/index.d.mts",
+        "default": "./dist/v3/workers/index.mjs"
+      },
+      "require": "./dist/v3/workers/index.js",
+      "types": "./dist/v3/workers/index.d.ts"
+    },
     "./package.json": "./package.json"
   },
   "typesVersions": {
@@ -71,25 +135,25 @@
     "@opentelemetry/sdk-trace-node": "^1.22.0",
     "@opentelemetry/semantic-conventions": "^1.22.0",
     "humanize-duration": "^3.27.3",
-    "socket.io": "^4.7.4",
-    "socket.io-client": "^4.7.4",
     "superjson": "^2.2.1",
     "ulidx": "^2.2.1",
     "zod": "3.22.3",
     "zod-error": "1.5.0",
-    "zod-validation-error": "^1.5.0"
+    "zod-validation-error": "^1.5.0",
+    "socket.io-client": "4.7.4"
   },
   "devDependencies": {
     "@trigger.dev/tsconfig": "workspace:*",
     "@trigger.dev/tsup": "workspace:*",
     "@types/humanize-duration": "^3.27.1",
     "@types/jest": "^29.5.3",
-    "@types/node": "^18",
+    "@types/node": "20.12.7",
     "jest": "^29.6.2",
     "rimraf": "^3.0.2",
     "ts-jest": "^29.1.1",
     "tsup": "^8.0.1",
-    "typescript": "^5.3.0"
+    "typescript": "^5.3.0",
+    "socket.io": "4.7.4"
   },
   "engines": {
     "node": ">=18.0.0"
