+++ packages/core/src/v3/dev/index.ts
@@ -1,0 +1,1 @@
+export { DevRuntimeManager } from "../runtime/devRuntimeManager";
