+++ packages/core/src/v3/prod/index.ts
@@ -1,0 +1,1 @@
+export { ProdRuntimeManager } from "../runtime/prodRuntimeManager";
