+++ packages/core/src/v3/workers/index.ts
@@ -1,0 +1,4 @@
+export { TaskExecutor, type TaskExecutorOptions } from "./taskExecutor";
+export { taskContextManager, TaskContextSpanProcessor } from "../tasks/taskContextManager";
+export type { RuntimeManager } from "../runtime/manager";
+export { PreciseWallClock as DurableClock } from "../clock/preciseWallClock";
